+++ src/lattice/iterator/mod.rs
@@ -5,6 +5,14 @@
 
 // TODO reduce identical code using private traits
 
+// TODO
+// - unify direction behavior (closure)
+// - bench
+// - possibly optimize next and next_back operation
+// - bound to the struct declaration
+// - more doc
+// - optimize other part of the code
+
 mod direction;
 mod link_canonical;
 mod point;
@@ -16,27 +24,31 @@ use std::{
 
 use rayon::iter::{
     plumbing::{bridge, Consumer, Producer, ProducerCallback, UnindexedConsumer},
-    IndexedParallelIterator, ParallelIterator,
+    IndexedParallelIterator, IntoParallelIterator, ParallelIterator,
 };
 #[cfg(feature = "serde-serialize")]
 use serde::{Deserialize, Serialize};
 use utils_lib::Getter;
 
 pub use self::direction::IteratorDirection;
-pub use self::link_canonical::IteratorLatticeLinkCanonical;
-pub use self::point::IteratorLatticePoint;
+pub use self::link_canonical::{IteratorLatticeLinkCanonical, ParIterLatticeLinkCanonical};
+pub use self::point::{IteratorLatticePoint, ParIterLatticePoint};
 use super::{IndexToElement, LatticeCyclic, LatticeElementToIndex, NumberOfLatticeElement};
 use crate::private::Sealed;
 
+/// Trait for generic implementation of [`Iterator`] for implementor of this trait.
+///
+/// It has a notion of dimension as it is link to the notion of lattice element.
+/// And a lattice takes a dimension.
 pub trait RandomAccessIterator<const D: usize>: Sealed {
+    /// Type of element return by the iterator
     type Item;
-    // type Iterator: Iterator<Item = &'a Self::Item>;
-    // type ParallelIterator: IndexedParallelIterator<Item = &'a Self::Item>;
 
-    /// returns the number of elements left in the iterator
+    /// Returns the number of elements left in the iterator.
     #[must_use]
     fn iter_len(&self) -> usize;
 
+    /// Increase the given front element by the given number .
     #[must_use]
     fn increase_element_by(
         lattice: &LatticeCyclic<D>,
@@ -44,6 +56,7 @@ pub trait RandomAccessIterator<const D: usize>: Sealed {
         advance_by: usize,
     ) -> IteratorElement<Self::Item>;
 
+    /// decrease the back element by a given number.
     #[must_use]
     fn decrease_element_by(
         lattice: &LatticeCyclic<D>,
@@ -101,6 +114,9 @@ impl<T> IteratorElement<T> {
         }
     }
 
+    /// Transform an index to an element mapping o to [`Self::FirstElement`] and
+    /// index - 1 to the closure. If the closure returns None [`Self::LastElement`] is
+    /// returned instead.
     #[inline]
     #[must_use]
     fn index_to_element<F: FnOnce(usize) -> Option<T>>(index: usize, closure: F) -> Self {
@@ -129,7 +145,7 @@ impl<T> IteratorElement<T> {
     }
 }
 
-#[doc(hidden)]
+//#[doc(hidden)]
 impl<T: Sealed> Sealed for IteratorElement<T> {}
 
 impl<T: Display> Display for IteratorElement<T> {
@@ -151,7 +167,7 @@ impl<T> Default for IteratorElement<T> {
     }
 }
 
-/// Returns [`Some`] in the case of [`IteratorElement::Element`], [`None`] otherwise
+/// Returns [`Some`] in the case of [`IteratorElement::Element`], [`None`] otherwise.
 impl<T> From<IteratorElement<T>> for Option<T> {
     #[inline]
     fn from(element: IteratorElement<T>) -> Self {
@@ -162,6 +178,9 @@ impl<T> From<IteratorElement<T>> for Option<T> {
     }
 }
 
+/// Index [`Self`] as 0 for the [`Self::FirstElement`],
+/// [`NumberOfLatticeElement::number_of_elements`] + 1 for [`Self::LastElement`]
+/// and index + 1 for any other element.
 impl<const D: usize, T: LatticeElementToIndex<D> + NumberOfLatticeElement<D>>
     LatticeElementToIndex<D> for IteratorElement<T>
 {
@@ -175,6 +194,7 @@ impl<const D: usize, T: LatticeElementToIndex<D> + NumberOfLatticeElement<D>>
     }
 }
 
+/// The number of element is the number of element of `T` + 2
 impl<const D: usize, T: LatticeElementToIndex<D> + NumberOfLatticeElement<D>>
     NumberOfLatticeElement<D> for IteratorElement<T>
 {
@@ -184,6 +204,12 @@ impl<const D: usize, T: LatticeElementToIndex<D> + NumberOfLatticeElement<D>>
     }
 }
 
+/// An iterator that track the front and the back in order to be able to implement
+/// [`DoubleEndedIterator`].
+///
+/// By itself it is not use a lot in the library it is used as a properties and use
+/// to track the front and the back. [`Iterator`] traits are not (yet ?) implemented
+/// on this type.
 #[derive(Getter, Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 struct DoubleEndedCounter<T> {
@@ -202,14 +228,24 @@ struct DoubleEndedCounter<T> {
 impl<T> DoubleEndedCounter<T> {
     /// Create a new [`Self`] with [`IteratorElement::FirstElement`] as `front` and
     /// [`IteratorElement::LastElement`] as `end`
+    /// # Example
+    /// ```ignore
+    /// use lattice_qcd_rs::lattice::iter::{DoubleEndedCounter,IteratorElement};
+    /// let counter = DoubleEndedCounter::<()>::new();
+    /// assert_eq!(counter.front(), IteratorElement::FirstElement);
+    /// assert_eq!(counter.end(), IteratorElement::LastElement);
+    /// ```
     pub const fn new() -> Self {
         Self {
             front: IteratorElement::FirstElement,
             end: IteratorElement::LastElement,
         }
     }
+
+    // possible with_first, with_last
 }
 
+/// It is [`Self::new`],
 impl<T> Default for DoubleEndedCounter<T> {
     #[inline]
     fn default() -> Self {
@@ -237,6 +273,30 @@ impl<T: Display> Display for DoubleEndedCounter<T> {
     }
 }
 
+/// Iterator over a lattice. This is a way to generate fast a ensemble of predetermine
+/// element without the need of being allocated for big collection. It is also possible
+/// to use a parallel version of the iterator (as explained below).
+///
+/// This struct is use for generically implementing [`Iterator`], [`DoubleEndedIterator`]
+/// [`ExactSizeIterator`] and [`FusedIterator`].
+///
+/// We can also easily transform to a parallel iterator [`LatticeParIter`] using
+/// [`Self::par_iter`], [`Self::as_par_iter`] or [`Self::as_par_iter_mut`].
+///
+/// It is only used
+/// for `T` = [`crate::lattice::LatticePoint`] (see [`IteratorLatticePoint`]) and
+/// for `T` = [`crate::lattice::LatticeLinkCanonical`] (see [`IteratorLatticeLinkCanonical`]).
+/// (Public) constructors only exist for these possibilities.
+///
+/// The iterators can be created from [`crate::lattice::LatticeCyclic::get_points`] and
+/// [`crate::lattice::LatticeCyclic::get_links`].
+///
+/// # Example
+/// TODO
+/// ```
+/// todo!()
+/// ```
+// TODO bound explanation
 #[derive(Getter, Debug, Clone, PartialEq)]
 pub struct LatticeIterator<'a, const D: usize, T> {
     /// ref to the lattice
@@ -252,33 +312,99 @@ pub struct LatticeIterator<'a, const D: usize, T> {
 impl<'a, const D: usize, T> LatticeIterator<'a, D, T> {
     /// create a new iterator with a ref to a given lattice, [`IteratorElement::FirstElement`]
     /// as the `front` and [`IteratorElement::LastElement`] as the `end`.
+    ///
+    /// This method is implemented only for T = [`crate::lattice::LatticePoint`]
+    /// or [`crate::lattice::LatticeLinkCanonical`].
+    ///
+    /// # Example
+    /// TODO
+    /// ```
+    /// use lattice_qcd_rs::lattice::{LatticeCyclic, LatticeIterator, LatticePoint};
+    /// use nalgebra::Vector4;
+    ///
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// let lattice = LatticeCyclic::<4>::new(1_f64, 3)?;
+    /// let mut iter = LatticeIterator::new(&lattice);
+    ///
+    /// assert_eq!(
+    ///     iter.next(),
+    ///     Some(LatticePoint::new(Vector4::new(0, 0, 0, 0)))
+    /// );
+    /// assert_eq!(
+    ///     iter.next(),
+    ///     Some(LatticePoint::new(Vector4::new(1, 0, 0, 0)))
+    /// );
+    /// assert_eq!(
+    ///     iter.next(),
+    ///     Some(LatticePoint::new(Vector4::new(2, 0, 0, 0)))
+    /// );
+    /// assert_eq!(
+    ///     iter.next(),
+    ///     Some(LatticePoint::new(Vector4::new(0, 1, 0, 0)))
+    /// );
+    /// // nth forward
+    /// assert_eq!(
+    ///     iter.nth(5),
+    ///     Some(LatticePoint::new(Vector4::new(0, 0, 1, 0)))
+    /// );
+    ///
+    /// // the iterator is double ended so we can poll back
+    /// assert_eq!(
+    ///     iter.next_back(),
+    ///     Some(LatticePoint::new(Vector4::new(2, 2, 2, 2)))
+    /// );
+    /// assert_eq!(
+    ///     iter.next_back(),
+    ///     Some(LatticePoint::new(Vector4::new(1, 2, 2, 2)))
+    /// );
+    /// assert_eq!(
+    ///     iter.next_back(),
+    ///     Some(LatticePoint::new(Vector4::new(0, 2, 2, 2)))
+    /// );
+    /// // we can also use `nth_back()`
+    /// assert_eq!(
+    ///     iter.nth_back(8),
+    ///     Some(LatticePoint::new(Vector4::new(0, 2, 1, 2)))
+    /// );
+    /// # Ok(())
+    /// # }
+    /// ```
+    // TODO
     #[must_use]
     #[inline]
-    pub const fn new(lattice: &'a LatticeCyclic<D>) -> Self {
+    pub const fn new(lattice: &'a LatticeCyclic<D>) -> Self
+    where
+        Self: RandomAccessIterator<D, Item = T>,
+        T: Clone,
+    {
         Self {
             lattice,
             counter: DoubleEndedCounter::new(),
         }
     }
 
+    /// Get the front element tracker.
     #[must_use]
     #[inline]
     const fn front(&self) -> &IteratorElement<T> {
         self.counter().front()
     }
 
+    /// Get the end element tracker.
     #[must_use]
     #[inline]
     const fn end(&self) -> &IteratorElement<T> {
         self.counter().end()
     }
 
+    /// Get a mutable reference on the front element tracker.
     #[must_use]
     #[inline]
     fn front_mut(&mut self) -> &mut IteratorElement<T> {
         self.counter_mut().front_mut()
     }
 
+    /// Get a mutable reference on the end element tracker.
     #[must_use]
     #[inline]
     fn end_mut(&mut self) -> &mut IteratorElement<T> {
@@ -286,6 +412,10 @@ impl<'a, const D: usize, T> LatticeIterator<'a, D, T> {
     }
 
     /// create a new iterator. The first [`LatticeIterator::next()`] will return `first_el`.
+    ///
+    /// This method is implemented only for T = [`crate::lattice::LatticePoint`]
+    /// or [`crate::lattice::LatticeLinkCanonical`].
+    ///
     /// # Example
     /// ```
     /// use lattice_qcd_rs::error::ImplementationError;
@@ -324,6 +454,7 @@ impl<'a, const D: usize, T> LatticeIterator<'a, D, T> {
     pub fn new_with_first_element(lattice: &'a LatticeCyclic<D>, first_el: T) -> Self
     where
         Self: RandomAccessIterator<D, Item = T>,
+        T: Clone,
     {
         Self {
             lattice,
@@ -334,31 +465,34 @@ impl<'a, const D: usize, T> LatticeIterator<'a, D, T> {
         }
     }
 
+    /// Convert the iterator into an [`IndexedParallelIterator`].
     #[must_use]
     #[inline]
-    pub fn par_iter(self) -> LatticeParIter<'a, D, T>
+    pub const fn par_iter(self) -> LatticeParIter<'a, D, T>
     where
         Self: RandomAccessIterator<D, Item = T>,
     {
-        self.into()
+        LatticeParIter::new(self)
     }
 
+    /// Take a reference of self and return a reference to an [`IndexedParallelIterator`].
+    /// This might not be very useful. Look instead at [`Self::as_par_iter_mut`].
+    #[allow(unsafe_code)]
     #[must_use]
     #[inline]
-    pub fn as_par_iter(&self) -> &LatticeParIter<'a, D, T>
-    where
-        Self: RandomAccessIterator<D, Item = T>,
-    {
-        self.as_ref()
+    pub const fn as_par_iter<'b>(&'b self) -> &'b LatticeParIter<'a, D, T> {
+        // SAFETY: the representation is transparent and the lifetime is not extended
+        unsafe { &*(self as *const Self).cast::<LatticeParIter<'a, D, T>>() }
     }
 
+    /// Take a mutable reference of self and return a mutable reference to an
+    /// [`IndexedParallelIterator`]
+    #[allow(unsafe_code)]
     #[must_use]
     #[inline]
-    pub fn as_par_iter_mut(&mut self) -> &mut LatticeParIter<'a, D, T>
-    where
-        Self: RandomAccessIterator<D, Item = T>,
-    {
-        self.as_mut()
+    pub fn as_par_iter_mut<'b>(&'b mut self) -> &'b mut LatticeParIter<'a, D, T> {
+        // SAFETY: the representation is transparent and the lifetime is not extended
+        unsafe { &mut *(self as *mut Self).cast::<LatticeParIter<'a, D, T>>() }
     }
 }
 
@@ -372,6 +506,7 @@ impl<'a, const D: usize, T: Display> Display for LatticeIterator<'a, D, T> {
     }
 }
 
+//#[doc(hidden)]
 impl<'a, const D: usize, T> Sealed for LatticeIterator<'a, D, T> {}
 
 impl<'a, const D: usize, T> AsRef<DoubleEndedCounter<T>> for LatticeIterator<'a, D, T> {
@@ -414,8 +549,6 @@ where
     T: LatticeElementToIndex<D> + NumberOfLatticeElement<D> + IndexToElement<D>,
 {
     type Item = T;
-    // type Iterator;
-    // type ParallelIterator;
 
     fn iter_len(&self) -> usize {
         // we use a saturating sub because the front could go further than the back
@@ -425,7 +558,6 @@ where
             .saturating_sub(self.end().size_position(self.lattice()))
     }
 
-    /// Increase an element by a given amount of step.
     fn increase_element_by(
         lattice: &LatticeCyclic<D>,
         front_element: &IteratorElement<Self::Item>,
@@ -434,22 +566,28 @@ where
         let index = match front_element {
             IteratorElement::FirstElement => 0,
             IteratorElement::Element(ref element) => element.to_index(lattice) + 1,
-            IteratorElement::LastElement => return IteratorElement::LastElement,
+            IteratorElement::LastElement => {
+                // early return
+                return IteratorElement::LastElement;
+            }
         };
+
         let new_index = index + advance_by;
         IteratorElement::index_to_element(new_index, |index| {
             Self::Item::index_to_element(lattice, index)
         })
     }
 
-    /// Decrease an element by a given amount of step.
     fn decrease_element_by(
         lattice: &LatticeCyclic<D>,
         end_element: &IteratorElement<Self::Item>,
         back_by: usize,
     ) -> IteratorElement<Self::Item> {
         let index = match end_element {
-            IteratorElement::FirstElement => return IteratorElement::FirstElement,
+            IteratorElement::FirstElement => {
+                // early return
+                return IteratorElement::FirstElement;
+            }
             IteratorElement::Element(ref element) => element.to_index(lattice) + 1,
             IteratorElement::LastElement => lattice.number_of_points() + 1,
         };
@@ -464,7 +602,7 @@ where
 /// TODO DOC
 impl<'a, const D: usize, T> Iterator for LatticeIterator<'a, D, T>
 where
-    LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
+    Self: RandomAccessIterator<D, Item = T>,
     T: Clone,
 {
     type Item = T;
@@ -534,7 +672,7 @@ where
 /// TODO DOC
 impl<'a, const D: usize, T> DoubleEndedIterator for LatticeIterator<'a, D, T>
 where
-    LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
+    Self: RandomAccessIterator<D, Item = T>,
     T: Clone,
 {
     #[inline]
@@ -580,14 +718,28 @@ where
 impl<'a, const D: usize, T> From<LatticeParIter<'a, D, T>> for LatticeIterator<'a, D, T> {
     #[inline]
     fn from(value: LatticeParIter<'a, D, T>) -> Self {
-        value.0
+        value.into_iterator()
     }
 }
 
 impl<'a, const D: usize, T> From<LatticeProducer<'a, D, T>> for LatticeIterator<'a, D, T> {
     #[inline]
     fn from(value: LatticeProducer<'a, D, T>) -> Self {
-        value.0
+        value.into_iterator()
+    }
+}
+
+impl<'a, const D: usize, T> IntoParallelIterator for LatticeIterator<'a, D, T>
+where
+    Self: RandomAccessIterator<D, Item = T>,
+    T: Clone + Send,
+{
+    type Iter = LatticeParIter<'a, D, T>;
+    type Item = <Self::Iter as ParallelIterator>::Item;
+
+    #[inline]
+    fn into_par_iter(self) -> Self::Iter {
+        self.par_iter()
     }
 }
 
@@ -595,6 +747,29 @@ impl<'a, const D: usize, T> From<LatticeProducer<'a, D, T>> for LatticeIterator<
 #[repr(transparent)]
 #[derive(Debug, Clone, PartialEq)]
 struct LatticeProducer<'a, const D: usize, T>(LatticeIterator<'a, D, T>);
+
+impl<'a, const D: usize, T> LatticeProducer<'a, D, T> {
+    /// Convert self into a [`LatticeIterator`]
+    #[inline]
+    #[must_use]
+    fn into_iterator(self) -> LatticeIterator<'a, D, T> {
+        self.0
+    }
+
+    /// Convert as a reference of [`LatticeIterator`]
+    #[inline]
+    #[must_use]
+    const fn as_iter(&self) -> &LatticeIterator<'a, D, T> {
+        &self.0
+    }
+
+    /// Convert as a mutable reference of [`LatticeIterator`]
+    #[inline]
+    #[must_use]
+    fn as_iter_mut(&mut self) -> &mut LatticeIterator<'a, D, T> {
+        &mut self.0
+    }
+}
 
 impl<'a, const D: usize, T> From<LatticeIterator<'a, D, T>> for LatticeProducer<'a, D, T> {
     #[inline]
@@ -620,83 +795,106 @@ where
 
     #[inline]
     fn into_iter(self) -> Self::IntoIter {
-        self.into()
+        self.into_iterator()
     }
 
     #[inline]
     fn split_at(self, index: usize) -> (Self, Self) {
         let splinting = Self::IntoIter::increase_element_by(
-            self.as_ref().lattice(),
-            self.as_ref().front(),
+            self.as_iter().lattice(),
+            self.as_iter().front(),
             index,
         );
         (
-            Self::IntoIter {
+            Self(Self::IntoIter {
                 lattice: self.0.lattice,
                 counter: DoubleEndedCounter {
                     front: self.0.counter.front,
                     end: splinting.clone(),
                 },
-            }
-            .into(),
-            Self::IntoIter {
+            }),
+            Self(Self::IntoIter {
                 lattice: self.0.lattice,
                 counter: DoubleEndedCounter {
                     front: splinting,
                     end: self.0.counter.end,
                 },
-            }
-            .into(),
+            }),
         )
     }
 }
 
 impl<'a, const D: usize, T> AsRef<LatticeIterator<'a, D, T>> for LatticeProducer<'a, D, T> {
+    #[inline]
     fn as_ref(&self) -> &LatticeIterator<'a, D, T> {
-        &self.0
+        self.as_iter()
     }
 }
 
 impl<'a, const D: usize, T> AsMut<LatticeIterator<'a, D, T>> for LatticeProducer<'a, D, T> {
+    #[inline]
     fn as_mut(&mut self) -> &mut LatticeIterator<'a, D, T> {
-        &mut self.0
+        self.as_iter_mut()
     }
 }
 
+impl<'a, const D: usize, T> IntoIterator for LatticeProducer<'a, D, T>
+where
+    LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
+    T: Clone,
+{
+    type Item = <Self::IntoIter as Iterator>::Item;
+    type IntoIter = LatticeIterator<'a, D, T>;
+
+    #[inline]
+    fn into_iter(self) -> Self::IntoIter {
+        self.into_iterator()
+    }
+}
+
+/// [`rayon::iter::ParallelIterator`] and [`rayon::iter::IndexedParallelIterator`]
+/// over a lattice. It is only used
+/// for `T` = [`crate::lattice::LatticePoint`] (see [`ParIterLatticePoint`]) and
+/// for `T` = [`crate::lattice::LatticeLinkCanonical`] (see [`ParIterLatticeLinkCanonical`]).
+/// (Public) constructors only exist for these possibilities using [`LatticeIterator::par_iter`].
+///
+/// It has a transparent representation containing a single field [`LatticeIterator`] allowing
+/// transmutation between the two. Though other crate should not rely on this representation.
+/// TODO more doc
 #[repr(transparent)]
 #[derive(Debug, Clone, PartialEq)]
 pub struct LatticeParIter<'a, const D: usize, T>(LatticeIterator<'a, D, T>);
 
-impl<'a, const D: usize, T> LatticeParIter<'a, D, T>
-where
-    LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
-    T: Clone + Send,
-{
-    #[must_use]
-    #[inline]
-    pub fn into_iter(self) -> LatticeIterator<'a, D, T>
-    where
-        LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
-    {
-        self.into()
+impl<'a, const D: usize, T> LatticeParIter<'a, D, T> {
+    // where
+    //LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
+    //T: Clone + Send,
+
+    /// create a new self with `iter` as the wrapped value
+    const fn new(iter: LatticeIterator<'a, D, T>) -> Self {
+        Self(iter)
     }
 
+    /// Convert the parallel iterator into an [`Iterator`]
     #[must_use]
     #[inline]
-    pub fn iter(&self) -> &LatticeIterator<'a, D, T>
-    where
-        LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
-    {
-        self.as_ref()
+    pub fn into_iterator(self) -> LatticeIterator<'a, D, T> {
+        self.0
     }
 
+    /// Take a reference of self and return a reference to an [`Iterator`].
+    /// This might not be very useful look instead at [`Self::as_iter_mut`]
     #[must_use]
     #[inline]
-    pub fn iter_mut(&mut self) -> &mut LatticeIterator<'a, D, T>
-    where
-        LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
-    {
-        self.as_mut()
+    pub const fn as_iter(&self) -> &LatticeIterator<'a, D, T> {
+        &self.0
+    }
+
+    /// Take a mutable reference of self and return a mutable reference to an [`Iterator`].
+    #[must_use]
+    #[inline]
+    pub fn as_iter_mut(&mut self) -> &mut LatticeIterator<'a, D, T> {
+        &mut self.0
     }
 }
 
@@ -771,7 +969,7 @@ where
 impl<'a, const D: usize, T> From<LatticeIterator<'a, D, T>> for LatticeParIter<'a, D, T> {
     #[inline]
     fn from(value: LatticeIterator<'a, D, T>) -> Self {
-        Self(value)
+        Self::new(value)
     }
 }
 
@@ -783,32 +981,44 @@ impl<'a, const D: usize, T> From<LatticeProducer<'a, D, T>> for LatticeParIter<'
 }
 
 impl<'a, const D: usize, T> AsRef<LatticeIterator<'a, D, T>> for LatticeParIter<'a, D, T> {
+    #[inline]
     fn as_ref(&self) -> &LatticeIterator<'a, D, T> {
-        &self.0
+        self.as_iter()
     }
 }
 
 impl<'a, const D: usize, T> AsMut<LatticeIterator<'a, D, T>> for LatticeParIter<'a, D, T> {
+    #[inline]
     fn as_mut(&mut self) -> &mut LatticeIterator<'a, D, T> {
-        &mut self.0
+        self.as_iter_mut()
     }
 }
 
 impl<'a, const D: usize, T> AsRef<LatticeParIter<'a, D, T>> for LatticeIterator<'a, D, T> {
     #[inline]
-    #[allow(unsafe_code)]
-    fn as_ref<'b>(&'b self) -> &'b LatticeParIter<'a, D, T> {
-        // SAFETY: the representation is transparent and the lifetime is not extended
-        unsafe { &*(self as *const Self).cast::<LatticeParIter<'a, D, T>>() }
+    fn as_ref(&self) -> &LatticeParIter<'a, D, T> {
+        self.as_par_iter()
     }
 }
 
 impl<'a, const D: usize, T> AsMut<LatticeParIter<'a, D, T>> for LatticeIterator<'a, D, T> {
     #[inline]
-    #[allow(unsafe_code)]
     fn as_mut(&mut self) -> &mut LatticeParIter<'a, D, T> {
-        // SAFETY: the representation is transparent and the lifetime is not extended
-        unsafe { &mut *(self as *mut Self).cast::<LatticeParIter<'a, D, T>>() }
+        self.as_par_iter_mut()
+    }
+}
+
+impl<'a, const D: usize, T> IntoIterator for LatticeParIter<'a, D, T>
+where
+    LatticeIterator<'a, D, T>: RandomAccessIterator<D, Item = T>,
+    T: Clone,
+{
+    type Item = <Self::IntoIter as Iterator>::Item;
+    type IntoIter = LatticeIterator<'a, D, T>;
+
+    #[inline]
+    fn into_iter(self) -> Self::IntoIter {
+        self.into_iterator()
     }
 }
 
@@ -816,7 +1026,9 @@ impl<'a, const D: usize, T> AsMut<LatticeParIter<'a, D, T>> for LatticeIterator<
 mod test {
     use std::error::Error;
 
-    use super::{IteratorElement, IteratorLatticeLinkCanonical, IteratorLatticePoint};
+    use super::{
+        DoubleEndedCounter, IteratorElement, IteratorLatticeLinkCanonical, IteratorLatticePoint,
+    };
     use crate::{
         error::ImplementationError,
         lattice::{DirectionEnum, LatticeCyclic, LatticeLinkCanonical, LatticePoint},
@@ -863,5 +1075,20 @@ mod test {
         );
 
         Ok(())
+    }
+
+    #[test]
+    fn double_ended_counter_default() {
+        assert_eq!(
+            DoubleEndedCounter::<()>::default(),
+            DoubleEndedCounter::<()>::new()
+        );
+        assert_eq!(
+            DoubleEndedCounter::<LatticeLinkCanonical<4>>::default(),
+            DoubleEndedCounter::<LatticeLinkCanonical<4>>::new()
+        );
+        let counter = DoubleEndedCounter::<()>::new();
+        assert_eq!(counter.front(), &IteratorElement::FirstElement);
+        assert_eq!(counter.end(), &IteratorElement::LastElement);
     }
 }